Changed Server.java so messages are no longer sent back to the client
Application seems to be in a somewhat working order. Users are able to join and send messages to each other. However I still need to implement the ability to register with a password, encrypt messages, store details in a database, restrict concurrent users to 3, use a semaphore to manage waiting clients, display timestamps, trigger a notification sound, send files, and potentially implement additional features. However we'll have to see about that as this coursework already seems to be taking more time than all the other 3 courseworks combined...